requirements.txt
Django>=4.2
gunicorn
mysqlclient
google-cloud-secret-manager



